refactor: move queryTimeline() into EventService
