+++ frontend/test/service/TimelineService.test.tsx
@@ -2,7 +2,7 @@ import { beforeEach, describe, it } from "@jest/globals";
 import IServerFacade from "@/service/IServerFacade";
 import { anything, instance, mock, verify, when } from "@typestrong/ts-mockito";
 import { AuthToken } from "@/model/model";
-import TimelineService, { TimelineQuery } from "@/service/TimelineService";
+import EventService, { TimelineQuery } from "@/service/EventService";
 
 const QUERY_W_NO_REQUIREMENTS: TimelineQuery = {
   requiredAttendees: []
@@ -14,12 +14,12 @@ const TOKEN: AuthToken = {
 
 describe("TimelineService", () => {
   let serverMock: IServerFacade;
-  let service: TimelineService;
+  let service: EventService;
 
   beforeEach(() => {
     serverMock = mock<IServerFacade>();
     const server = instance(serverMock);
-    service = new TimelineService(server);
+    service = new EventService(server);
 
     when(serverMock.queryTimeline(anything(), anything())).thenResolve(
       { events: [] }
